fixes get logs my family id endpoint by moving into different router
