+++ api/guestLogs/logsRouter.js
@@ -1,3 +1,4 @@
+// The get all logs by family id is located in family router
 const express = require('express');
 
 const Logs = require('./logsModel');
@@ -30,21 +31,6 @@ router.get('/:id', function (req, res) {
     });
 });
 
-//get all logs by family id
-router.get('families/:id', function (req, res) {
-  const family_id = String(req.params.id);
-  Logs.findByFamilyId(family_id)
-    .then((logs) => {
-      if (logs) {
-        res.status(200).json(logs);
-      } else {
-        res.status(404).json({ error: 'logs NotFound' });
-      }
-    })
-    .catch((err) => {
-      res.status(500).json({ error: err.message });
-    });
-});
 router.post('/', async (req, res) => {
   const logs = req.body;
   if (logs) {
